Commit the junk in my tree before I go back to school:
        * ltconfig
        * ltmain.sh: downgraded to libtool 1.0f at Jay's suggestion

        * app/gimage.c: Fix for the "merge to alpha" bug from Nathan
        Summers

        * app/text_tool.c: Fix for the font spec bug from Nathan
        Summers

        * plug-ins/maze/maze.c
        * plug-ins/rotate/rotate.c: updated from registry

        * plug-ins/blur/blur.c
        * plug-ins/randomize/randomize.c
        * plug-ins/vpropagate/vpropagate.c
        * plug-ins/warp/warp.c: more portability fixes for DEC OSF1

-Yosh

+++ plug-ins/common/randomize.c
@@ -474,7 +474,7 @@ randomize_prepare_row(GPixelRgn *pixel_rgn, guchar *data, int x, int y, int w)
  *  Fill in edge pixels
  */
     for (b = 0; b < pixel_rgn->bpp; b++) {
-	data[-pixel_rgn->bpp + b] = data[b];
+	data[-(gint)pixel_rgn->bpp + b] = data[b];
 	data[w * pixel_rgn->bpp + b] = data[(w - 1) * pixel_rgn->bpp + b];
     }
 }
@@ -494,7 +494,7 @@ randomize_prepare_row(GPixelRgn *pixel_rgn, guchar *data, int x, int y, int w)
  *  Fill in edge pixels \
  */ \
     for (b = 0; b < (pixel_rgn)->bpp; b++) { \
-	data[-(pixel_rgn)->bpp + b] = data[b]; \
+	data[-(gint)(pixel_rgn)->bpp + b] = data[b]; \
 	data[w * (pixel_rgn)->bpp + b] = data[(w - 1) * (pixel_rgn)->bpp + b]; \
     } \
 }
