+++ plug-ins/warp/warp.c
@@ -658,7 +658,7 @@ diff_prepare_row (GPixelRgn *pixel_rgn,
   /*  Fill in edge pixels  */
   for (b = 0; b < pixel_rgn->bpp; b++)
     {
-      data[-pixel_rgn->bpp + b] = data[b];
+      data[-(gint)pixel_rgn->bpp + b] = data[b];
       data[w * pixel_rgn->bpp + b] = data[(w - 1) * pixel_rgn->bpp + b];
     }
 }
